+++ gui/ui/actions.c
@@ -745,7 +745,7 @@ void uiCurr(void)
         if (curr) {
             uiSetFile(curr->path, curr->name, STREAMTYPE_FILE);
             guiInfo.PlaylistNext = False;
-            guiInfo.Track = (int)listMgr(PLAYLIST_ITEM_GET_POS, curr);
+            guiInfo.Track = (intptr_t)listMgr(PLAYLIST_ITEM_GET_POS, curr);
             break;
         }
 
@@ -815,7 +815,7 @@ void uiPrev(void)
         if (prev) {
             uiSetFile(prev->path, prev->name, STREAMTYPE_FILE);
             guiInfo.PlaylistNext = !guiInfo.Playing;
-            guiInfo.Track = (int)listMgr(PLAYLIST_ITEM_GET_POS, prev);
+            guiInfo.Track = (intptr_t)listMgr(PLAYLIST_ITEM_GET_POS, prev);
             break;
         }
 
@@ -883,7 +883,7 @@ void uiNext(void)
         if (next) {
             uiSetFile(next->path, next->name, STREAMTYPE_FILE);
             guiInfo.PlaylistNext = !guiInfo.Playing;
-            guiInfo.Track = (int)listMgr(PLAYLIST_ITEM_GET_POS, next);
+            guiInfo.Track = (intptr_t)listMgr(PLAYLIST_ITEM_GET_POS, next);
             break;
         }
